BE: 랭킹 동점 순위 처리 방법 수정

Closes #110

See merge request s11-bigdata-dist-sub1/S11P21B306!153

+++ backend/src/main/java/com/ssafy/newstock/memberstocks/controller/response/AssetInfoResponse.java
@@ -12,11 +12,11 @@ public class AssetInfoResponse {
     private Long deposit;
     private Long profitAndLoss;
     private String ROI;
-    private Long rank;
+    private Integer rank;
     private String rankSaveTime;
 
 
-    public AssetInfoResponse(String nickname,Long totalPrice,Long deposit, Long profitAndLoss, String ROI, Long rank,String rankSaveTime){
+    public AssetInfoResponse(String nickname,Long totalPrice,Long deposit, Long profitAndLoss, String ROI, Integer rank,String rankSaveTime){
         this.nickname=nickname;
         this.totalPrice=totalPrice;
         this.deposit=deposit;
